Build wheels with Py 3.5, not 3.9
PEP 652 clarifies that the stable ABI is officially only backward-compatible, not forward-compatible. I was not aware of this when I wrote this wheels.yml generator script initially. It's not currently a problem (we unit test on all supported Python versions on all platforms, so we know that nothing is currently broken), but this change might be helpful for the future.